requirements.txt
openai-whisper
pyaudio==0.2.13
soundfile==0.12.1
pyttsx3
pyautogui
langchain
langchain-ollama
ollama